update: add filter in skill map results graph
